refactor: move transaction requests to its own module
